requirements.txt
pillow
scapy

#Optional Build Tools
pyinstaller
